Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,10 +2,10 @@
 #include <stdio.h>
 int main()
 {
-    int x,y, product=0;
-    scanf("%d %d", &x,&y);
+    float x,y, product=0;
+    scanf("%f %f", &x,&y);
     printf("");
     product=x*y;
-    printf("Product: %d",product);
+    printf("Product: %.2f",product);
     return 0;
 }
